sbi: Move base to its own file
Continue to reduce clutter in sbi.rs.
