+++ sbi/src/sbi.rs
@@ -13,6 +13,9 @@ mod error;
 pub use error::*;
 mod function;
 pub use function::*;
+// The Base SBI extension
+mod base;
+pub use base::*;
 // The reset SBI extension
 mod reset;
 pub use reset::*;
@@ -28,66 +31,6 @@ pub mod api;
 
 #[cfg(all(target_arch = "riscv64", target_os = "none"))]
 use core::arch::asm;
-
-/// Functions defined for the Base extension
-#[derive(Clone, Copy)]
-pub enum BaseFunction {
-    /// Returns the implemented version of the SBI standard.
-    GetSpecificationVersion,
-    /// Returns the ID of the SBI implementation.
-    GetImplementationID,
-    /// Returns the version of this SBI implementation.
-    GetImplementationVersion,
-    /// Checks if the given SBI extension is supported.
-    ProbeSbiExtension(u64),
-    /// Returns the vendor that produced this machine(`mvendorid`).
-    GetMachineVendorID,
-    /// Returns the architecture implementation ID this machine(`marchid`).
-    GetMachineArchitectureID,
-    /// Returns the implementation ID of this machine(`mimpid`).
-    GetMachineImplementationID,
-}
-
-impl BaseFunction {
-    /// Attempts to parse `Self` from the passed in `a0-a7`.
-    fn from_regs(args: &[u64]) -> Result<Self> {
-        use BaseFunction::*;
-
-        match args[6] {
-            0 => Ok(GetSpecificationVersion),
-            1 => Ok(GetImplementationID),
-            2 => Ok(GetImplementationVersion),
-            3 => Ok(ProbeSbiExtension(args[0])),
-            4 => Ok(GetMachineVendorID),
-            5 => Ok(GetMachineArchitectureID),
-            6 => Ok(GetMachineImplementationID),
-            _ => Err(Error::NotSupported),
-        }
-    }
-}
-
-impl SbiFunction for BaseFunction {
-    fn a6(&self) -> u64 {
-        use BaseFunction::*;
-        match self {
-            GetSpecificationVersion => 0,
-            GetImplementationID => 1,
-            GetImplementationVersion => 2,
-            ProbeSbiExtension(_) => 3,
-            GetMachineVendorID => 4,
-            GetMachineArchitectureID => 5,
-            GetMachineImplementationID => 6,
-        }
-    }
-
-    fn a0(&self) -> u64 {
-        use BaseFunction::*;
-        match self {
-            ProbeSbiExtension(ext) => *ext,
-            _ => 0,
-        }
-    }
-}
 
 /// Functions provided by the attestation extension.
 #[derive(Copy, Clone)]
